Move GameBoard related models to their own folder
